+++ CHANGELOG.md
@@ -51,7 +51,12 @@
 - Modules
   - General
       - Updated module names along with some descriptions and issue names.
-  - Removed
+- Issues
+    - Added attribute ```remarks``` holding a ```Hash``` of remarks about
+        that issue with the entity which made the remark as _key_ and an ```Array```
+        of remarks as _value_.
+    - Added method ```#add_remark```, allowing new remarks to be added to the ```Issue```.
+- Removed
       - ```xss_uri``` compatibility module.
 - Added
   - Buffer classes